backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0
pydantic>=2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
python-multipart>=0.0.5
httpx>=0.24.0
PyPDF2>=3.0.1
python-docx>=1.1.0
